Final cleanup pass: includes, tests, README
Include cleanup:
- Remove 17 unused headers from SimulationLoop.cpp: all system types
  (AbilitySystem, MinionWaveSystem, VFXRenderer, etc.) are forward-
  declared in GameSystems.h (via SimulationLoop.h); no method calls on
  those pointers in SimulationLoop.cpp itself

Test suite:
- Register test_fog in CMakeLists.txt (FogSystem CPU-only tests; 7 assertions)
- Fix test_fog.cpp: std::clamp(int, int, uint32_t-1) type mismatch
- CTest now runs 12/12 tests (was 11)
- 8 remaining unregistered tests (jungle, minion, structure, networking,
  flowfield, snapshot, determinism) reference modules not yet implemented;
  left as-is per 'do not remove any test' rule

Orphaned header audit:
- No truly orphaned headers: all 79 headers without matching .cpp are
  either header-only (ThreadPool, FrameAllocator, Frustum, SplineUtil...)
  or implemented in a merged .cpp (SoundHandle.h -> AudioEngine.cpp,
  BloomPass.h -> PostProcessPasses.cpp, etc.)

Dead code audit:
- No #if 0 blocks found anywhere in src/
- All multi-line // comment blocks are documentation, not dead code

README:
- Update Project Structure section: shows merged file groupings,
  3 docs files with descriptions, full 19-file test list

All 12/12 tests pass.

+++ tests/test_fog.cpp
@@ -160,7 +160,7 @@ void test_soft_falloff() {
   float edgeX = 100.0f + edgeDist;
   int ex =
       static_cast<int>((edgeX / FogSystem::WORLD_SIZE) * FogSystem::MAP_SIZE);
-  ex = std::clamp(ex, 0, FogSystem::MAP_SIZE - 1);
+  ex = std::clamp(ex, 0, static_cast<int>(FogSystem::MAP_SIZE) - 1);
   uint8_t edgeVal = fog.getVisionBuffer()[cx * FogSystem::MAP_SIZE + ex];
   assert(edgeVal < 255);
